Migrate auto-run scripts to ScriptModule
